+++ CHANGELOG.rst
@@ -17,7 +17,9 @@ dev
 
 - Headers marked as `sensitive` will no longer log their value at DEBUG level. Instead a placeholder value of `SENSITIVE_REDACTED` is logged.
 - Fixed perfect match missed for headers with empty values.
+- Restricted variable integer decoding to uint32 to prevent run-away computation. With thanks to `Hiroki Nishino`_.
 
+.. _Hiroki Nishino: https://github.com/tawAsh1
 
 4.1.0 (2025-01-22)
 ------------------